Allowing only owner of the image to delete it

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -151,9 +151,18 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+        User user = (User)session.getAttribute("loggeduser");
+        String logger = user.getUsername();
+        if(imageService.checkowner(imageId,logger)){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError",error);
+            return "images/image";
+        }
     }
